- Fixed issue with meal dishes not being properly marked as deleted on menu save adding them to a list in session to be updated when the menu changes are saved. - Other minor changes.

+++ trunk/GreenRiverCommons/src/com/greenriver/commons/DateRange.java
@@ -99,8 +99,8 @@ public class DateRange implements Comparable<DateRange>, Cloneable, Serializable
 
     public DateRange(Date min, Date max) {
         this(
-                min != null ? min.getTime() : null,
-                max != null ? max.getTime() : null);
+                (Long)(min != null ? min.getTime() : null),
+                (Long)(max != null ? max.getTime() : null));
     }
 
     public DateRange(Long min, Long max) {
